Ignore empty lines
As in "don't list them in the problems section at the bottom of the
page".

Also call the unparsed lines section "Ignored" rather than "Lines we
failed to parse".

"Failure" indicated there was a problem, while the truth might just as
well be that we're looking at an unimplemented already-planned
feature, or that it's just junk at the start or end of the thread
dump.