Refactor stack frame handling
The internal representation is now just an array of stack frame function names.

For places where we want to render this, we now have a method called
"decorateStackFrames" that will add " at " to the start of all stack
frames, and replace an empty stack with a single frame containing the
EMPTY_STACK pattern.