Ignore hard WAITING for notification case
When a thread is WAITING for notification but it won't tell us on what
and it's holding more than one lock. Then we could guess on the first
lock, but that would be just a guess so we just ignore this case until
we get an example where this actually happens.